Time: 2 ms (57.64%) | Memory: 45 MB (43.60%) - LeetSync

+++ 152-maximum-product-subarray/maximum-product-subarray.java
@@ -1,10 +1,10 @@
 class Solution {
     public int maxProduct(int[] nums) {
-        double n = nums.length;
-        double max = nums[0],currMax = nums[0],currMin = nums[0];
+        int n = nums.length;
+        int max = nums[0],currMax = nums[0],currMin = nums[0];
 
         for(int i=1;i<n;i++){
-            double temp = currMax;
+            int temp = currMax;
             currMax = Math.max(nums[i], Math.max(nums[i]*currMax,nums[i]*currMin));
             currMin = Math.min(nums[i], Math.min(nums[i]*currMin,nums[i]*temp));
 
@@ -12,6 +12,6 @@ class Solution {
         }
 
         
-        return (int)max;
+        return max;
     }
 }
